Fixed the warning about (const void*)element.Offset

+++ Timefall/src/Platform/OpenGL/OpenGLVertexArray.cpp
@@ -58,7 +58,7 @@ namespace Timefall
 		for (const BufferElement& element : layout)
 		{
 			glEnableVertexAttribArray(index);
-			glVertexAttribPointer(index, element.GetComponentCount(), ShaderDataTypeToGLBaseType(element.Type), element.Normalized ? GL_TRUE : GL_FALSE, layout.GetStride(), (const void*)element.Offset);
+			glVertexAttribPointer(index, element.GetComponentCount(), ShaderDataTypeToGLBaseType(element.Type), element.Normalized ? GL_TRUE : GL_FALSE, layout.GetStride(), (const void*)(intptr_t)element.Offset);
 			index++;
 		}
